Bug fix in the installation instructions and in the option protections

See merge request akantu/akantu!71

+++ python/akantu/__init__.py
@@ -63,7 +63,7 @@ class AkantuSparseMatrix (_aka_sparse.coo_matrix):
             data = _aka_np.concatenate((data, data_sup))
 
         _aka_sparse.coo_matrix.__init__(
-            self, (data, (row, col)), shape=(sz, sz))
+            self, (data, (row, col)), shape=(sz, sz), dtype=data.dtype)
 
 
 FromStress = _py11_akantu.FromHigherDim
@@ -71,4 +71,3 @@ FromTraction = _py11_akantu.FromSameDim
 _py11_akantu.__initialize()
 
 __version__ = _py11_akantu.getVersion()
-
